Address code review feedback: Refactor to DRY concern and simplify error handling
